+++ lib/models/memory_analytics.freezed.dart
@@ -1,0 +1,382 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'memory_analytics.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+MemoryAnalytics _$MemoryAnalyticsFromJson(Map<String, dynamic> json) {
+  return _MemoryAnalytics.fromJson(json);
+}
+
+/// @nodoc
+mixin _$MemoryAnalytics {
+  int get retentionRatePercent => throw _privateConstructorUsedError;
+  int get totalWordsLearned => throw _privateConstructorUsedError;
+  int get masteredCount => throw _privateConstructorUsedError;
+  int get learningCount => throw _privateConstructorUsedError;
+  int get dueCount => throw _privateConstructorUsedError;
+  int get a1Count => throw _privateConstructorUsedError;
+  int get a2Count => throw _privateConstructorUsedError;
+  int get b1Count => throw _privateConstructorUsedError;
+  int get b2Count => throw _privateConstructorUsedError;
+  List<int> get weeklyXpList => throw _privateConstructorUsedError;
+
+  /// Serializes this MemoryAnalytics to a JSON map.
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+
+  /// Create a copy of MemoryAnalytics
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $MemoryAnalyticsCopyWith<MemoryAnalytics> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $MemoryAnalyticsCopyWith<$Res> {
+  factory $MemoryAnalyticsCopyWith(
+          MemoryAnalytics value, $Res Function(MemoryAnalytics) then) =
+      _$MemoryAnalyticsCopyWithImpl<$Res, MemoryAnalytics>;
+  @useResult
+  $Res call(
+      {int retentionRatePercent,
+      int totalWordsLearned,
+      int masteredCount,
+      int learningCount,
+      int dueCount,
+      int a1Count,
+      int a2Count,
+      int b1Count,
+      int b2Count,
+      List<int> weeklyXpList});
+}
+
+/// @nodoc
+class _$MemoryAnalyticsCopyWithImpl<$Res, $Val extends MemoryAnalytics>
+    implements $MemoryAnalyticsCopyWith<$Res> {
+  _$MemoryAnalyticsCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of MemoryAnalytics
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? retentionRatePercent = null,
+    Object? totalWordsLearned = null,
+    Object? masteredCount = null,
+    Object? learningCount = null,
+    Object? dueCount = null,
+    Object? a1Count = null,
+    Object? a2Count = null,
+    Object? b1Count = null,
+    Object? b2Count = null,
+    Object? weeklyXpList = null,
+  }) {
+    return _then(_value.copyWith(
+      retentionRatePercent: null == retentionRatePercent
+          ? _value.retentionRatePercent
+          : retentionRatePercent // ignore: cast_nullable_to_non_nullable
+              as int,
+      totalWordsLearned: null == totalWordsLearned
+          ? _value.totalWordsLearned
+          : totalWordsLearned // ignore: cast_nullable_to_non_nullable
+              as int,
+      masteredCount: null == masteredCount
+          ? _value.masteredCount
+          : masteredCount // ignore: cast_nullable_to_non_nullable
+              as int,
+      learningCount: null == learningCount
+          ? _value.learningCount
+          : learningCount // ignore: cast_nullable_to_non_nullable
+              as int,
+      dueCount: null == dueCount
+          ? _value.dueCount
+          : dueCount // ignore: cast_nullable_to_non_nullable
+              as int,
+      a1Count: null == a1Count
+          ? _value.a1Count
+          : a1Count // ignore: cast_nullable_to_non_nullable
+              as int,
+      a2Count: null == a2Count
+          ? _value.a2Count
+          : a2Count // ignore: cast_nullable_to_non_nullable
+              as int,
+      b1Count: null == b1Count
+          ? _value.b1Count
+          : b1Count // ignore: cast_nullable_to_non_nullable
+              as int,
+      b2Count: null == b2Count
+          ? _value.b2Count
+          : b2Count // ignore: cast_nullable_to_non_nullable
+              as int,
+      weeklyXpList: null == weeklyXpList
+          ? _value.weeklyXpList
+          : weeklyXpList // ignore: cast_nullable_to_non_nullable
+              as List<int>,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$MemoryAnalyticsImplCopyWith<$Res>
+    implements $MemoryAnalyticsCopyWith<$Res> {
+  factory _$$MemoryAnalyticsImplCopyWith(_$MemoryAnalyticsImpl value,
+          $Res Function(_$MemoryAnalyticsImpl) then) =
+      __$$MemoryAnalyticsImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call(
+      {int retentionRatePercent,
+      int totalWordsLearned,
+      int masteredCount,
+      int learningCount,
+      int dueCount,
+      int a1Count,
+      int a2Count,
+      int b1Count,
+      int b2Count,
+      List<int> weeklyXpList});
+}
+
+/// @nodoc
+class __$$MemoryAnalyticsImplCopyWithImpl<$Res>
+    extends _$MemoryAnalyticsCopyWithImpl<$Res, _$MemoryAnalyticsImpl>
+    implements _$$MemoryAnalyticsImplCopyWith<$Res> {
+  __$$MemoryAnalyticsImplCopyWithImpl(
+      _$MemoryAnalyticsImpl _value, $Res Function(_$MemoryAnalyticsImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of MemoryAnalytics
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? retentionRatePercent = null,
+    Object? totalWordsLearned = null,
+    Object? masteredCount = null,
+    Object? learningCount = null,
+    Object? dueCount = null,
+    Object? a1Count = null,
+    Object? a2Count = null,
+    Object? b1Count = null,
+    Object? b2Count = null,
+    Object? weeklyXpList = null,
+  }) {
+    return _then(_$MemoryAnalyticsImpl(
+      retentionRatePercent: null == retentionRatePercent
+          ? _value.retentionRatePercent
+          : retentionRatePercent // ignore: cast_nullable_to_non_nullable
+              as int,
+      totalWordsLearned: null == totalWordsLearned
+          ? _value.totalWordsLearned
+          : totalWordsLearned // ignore: cast_nullable_to_non_nullable
+              as int,
+      masteredCount: null == masteredCount
+          ? _value.masteredCount
+          : masteredCount // ignore: cast_nullable_to_non_nullable
+              as int,
+      learningCount: null == learningCount
+          ? _value.learningCount
+          : learningCount // ignore: cast_nullable_to_non_nullable
+              as int,
+      dueCount: null == dueCount
+          ? _value.dueCount
+          : dueCount // ignore: cast_nullable_to_non_nullable
+              as int,
+      a1Count: null == a1Count
+          ? _value.a1Count
+          : a1Count // ignore: cast_nullable_to_non_nullable
+              as int,
+      a2Count: null == a2Count
+          ? _value.a2Count
+          : a2Count // ignore: cast_nullable_to_non_nullable
+              as int,
+      b1Count: null == b1Count
+          ? _value.b1Count
+          : b1Count // ignore: cast_nullable_to_non_nullable
+              as int,
+      b2Count: null == b2Count
+          ? _value.b2Count
+          : b2Count // ignore: cast_nullable_to_non_nullable
+              as int,
+      weeklyXpList: null == weeklyXpList
+          ? _value._weeklyXpList
+          : weeklyXpList // ignore: cast_nullable_to_non_nullable
+              as List<int>,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$MemoryAnalyticsImpl implements _MemoryAnalytics {
+  const _$MemoryAnalyticsImpl(
+      {this.retentionRatePercent = 90,
+      this.totalWordsLearned = 0,
+      this.masteredCount = 0,
+      this.learningCount = 0,
+      this.dueCount = 0,
+      this.a1Count = 0,
+      this.a2Count = 0,
+      this.b1Count = 0,
+      this.b2Count = 0,
+      final List<int> weeklyXpList = const [15, 25, 40, 30, 50, 20, 35]})
+      : _weeklyXpList = weeklyXpList;
+
+  factory _$MemoryAnalyticsImpl.fromJson(Map<String, dynamic> json) =>
+      _$$MemoryAnalyticsImplFromJson(json);
+
+  @override
+  @JsonKey()
+  final int retentionRatePercent;
+  @override
+  @JsonKey()
+  final int totalWordsLearned;
+  @override
+  @JsonKey()
+  final int masteredCount;
+  @override
+  @JsonKey()
+  final int learningCount;
+  @override
+  @JsonKey()
+  final int dueCount;
+  @override
+  @JsonKey()
+  final int a1Count;
+  @override
+  @JsonKey()
+  final int a2Count;
+  @override
+  @JsonKey()
+  final int b1Count;
+  @override
+  @JsonKey()
+  final int b2Count;
+  final List<int> _weeklyXpList;
+  @override
+  @JsonKey()
+  List<int> get weeklyXpList {
+    if (_weeklyXpList is EqualUnmodifiableListView) return _weeklyXpList;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_weeklyXpList);
+  }
+
+  @override
+  String toString() {
+    return 'MemoryAnalytics(retentionRatePercent: $retentionRatePercent, totalWordsLearned: $totalWordsLearned, masteredCount: $masteredCount, learningCount: $learningCount, dueCount: $dueCount, a1Count: $a1Count, a2Count: $a2Count, b1Count: $b1Count, b2Count: $b2Count, weeklyXpList: $weeklyXpList)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$MemoryAnalyticsImpl &&
+            (identical(other.retentionRatePercent, retentionRatePercent) ||
+                other.retentionRatePercent == retentionRatePercent) &&
+            (identical(other.totalWordsLearned, totalWordsLearned) ||
+                other.totalWordsLearned == totalWordsLearned) &&
+            (identical(other.masteredCount, masteredCount) ||
+                other.masteredCount == masteredCount) &&
+            (identical(other.learningCount, learningCount) ||
+                other.learningCount == learningCount) &&
+            (identical(other.dueCount, dueCount) ||
+                other.dueCount == dueCount) &&
+            (identical(other.a1Count, a1Count) || other.a1Count == a1Count) &&
+            (identical(other.a2Count, a2Count) || other.a2Count == a2Count) &&
+            (identical(other.b1Count, b1Count) || other.b1Count == b1Count) &&
+            (identical(other.b2Count, b2Count) || other.b2Count == b2Count) &&
+            const DeepCollectionEquality()
+                .equals(other._weeklyXpList, _weeklyXpList));
+  }
+
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  int get hashCode => Object.hash(
+      runtimeType,
+      retentionRatePercent,
+      totalWordsLearned,
+      masteredCount,
+      learningCount,
+      dueCount,
+      a1Count,
+      a2Count,
+      b1Count,
+      b2Count,
+      const DeepCollectionEquality().hash(_weeklyXpList));
+
+  /// Create a copy of MemoryAnalytics
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$MemoryAnalyticsImplCopyWith<_$MemoryAnalyticsImpl> get copyWith =>
+      __$$MemoryAnalyticsImplCopyWithImpl<_$MemoryAnalyticsImpl>(
+          this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$MemoryAnalyticsImplToJson(
+      this,
+    );
+  }
+}
+
+abstract class _MemoryAnalytics implements MemoryAnalytics {
+  const factory _MemoryAnalytics(
+      {final int retentionRatePercent,
+      final int totalWordsLearned,
+      final int masteredCount,
+      final int learningCount,
+      final int dueCount,
+      final int a1Count,
+      final int a2Count,
+      final int b1Count,
+      final int b2Count,
+      final List<int> weeklyXpList}) = _$MemoryAnalyticsImpl;
+
+  factory _MemoryAnalytics.fromJson(Map<String, dynamic> json) =
+      _$MemoryAnalyticsImpl.fromJson;
+
+  @override
+  int get retentionRatePercent;
+  @override
+  int get totalWordsLearned;
+  @override
+  int get masteredCount;
+  @override
+  int get learningCount;
+  @override
+  int get dueCount;
+  @override
+  int get a1Count;
+  @override
+  int get a2Count;
+  @override
+  int get b1Count;
+  @override
+  int get b2Count;
+  @override
+  List<int> get weeklyXpList;
+
+  /// Create a copy of MemoryAnalytics
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$MemoryAnalyticsImplCopyWith<_$MemoryAnalyticsImpl> get copyWith =>
+      throw _privateConstructorUsedError;
+}